Decouple download artifact helpers from api/_shared.ts
